+++ papi_sdk/tests/test_base.py
@@ -1,9 +1,10 @@
 import unittest
+import uuid
 
 from papi_sdk import APIv3
 
 
 class BaseTest(unittest.TestCase):
-    client = APIv3(key=f"1234:str(uuid.uuid4())")
+    client = APIv3(key=f"1234:{str(uuid.uuid4())}")
     status_ok = "ok"
     status_error = "error"